created resume button and logic for database to support join game and resume

+++ db/Game_users.js
@@ -7,21 +7,15 @@ class Game_users extends ActiveRecord {
   game = -1;
   username = '';
 
-  static async joinedAndNotStarted(userId, gameId) {
+  // userId, gameId -> boolean
+  static async isJoined(userId, gameId) {
     let query = `SELECT COUNT(*) FROM game_users 
       WHERE user_id = ${userId} AND game_id = ${gameId}`
-    let query2 = `SELECT started FROM games WHERE id = ${gameId}`
 
     let { count } = await db.one(query)
     let joined = parseInt(count);
-    //console.log("joined: ", joined);
-    let { started } = await db.one(query2)
-    //console.log("started", started);
-
-    if (started == 0) started = false
-    else started = true
-
-    return joined && !started
+    
+    return joined
   }
 
   // convert to player cards status change
@@ -30,10 +24,14 @@ class Game_users extends ActiveRecord {
 
   // userId, gameId -> void
   static async joinGame(gameId, userId) {
+    // DOUBLE CHECK if already joined
+    let isJoined = await Game_users.isJoined(userId, gameId);
+    if (isJoined)
+      return 
     //get current players and increment for player_num
     let { count: numPlayers } = await Games.getNumPlayers(gameId);
-    console.log(numPlayers);
 
+    // TODO numPlayers logic is broken BRB 5 minutes
     let query = `INSERT INTO game_users(game_id, user_id, player_num, winner)
       VALUES(${gameId}, ${userId}, ${++numPlayers}, 0)`;
 
@@ -41,10 +39,19 @@ class Game_users extends ActiveRecord {
   }
 
   // userId, gameId -> void
-  static leaveGame(gameId, userId) {
-    query = `DELETE FROM game_users
-      WHERE user_id = ${userId} AND game_id = ${gameId}`
-    db.none(query);
+  static async leaveGame(gameId, userId) {
+    let started = await Games.isStarted(gameId);
+
+    if (!started) {
+      let query = `DELETE FROM game_users
+        WHERE user_id = ${userId} AND game_id = ${gameId}
+        RETURNING player_num`;
+      let playerNum = await db.one(query);
+      // DECREMENT game_users that are > user we deleted
+      //query = `UPDATE game_users SET
+    } else {
+      console.log('ERROR leaveGame: cant leave game already started');
+    }
   }
 
   drawCard() {
